Atom and RSS api format for contest Fixed parsing events and statistics Added algotester Fixed tco stage stadings Messaging Manual linking accounts

+++ legacy/update.php
@@ -386,7 +386,9 @@
                 echo "\t<span style='padding-left: 50px'>" .
                     "{$contest['title']} [{$contest['key']}] <{$contest['url']}></span><br>\n";
             }
-            $db->delete("clist_contest", $query);
+            if (!DEBUG) {
+                $db->delete("clist_contest", $query);
+            }
         }
     }
     logmsg();
